+++ test/test_utils/test_hpbandster2smac.py
@@ -17,10 +17,10 @@ class TestHpbandster2Smac(unittest.TestCase):
         """ Having ints and bools as categoricals """
         try:
             hpbandster2smac = HpBandSter2SMAC()
-            hpbandster2smac.convert(self.path_to_result_mixed_categorical_json)
-            hpbandster2smac.convert(self.path_to_result_mixed_categorical_pcs)
+            hpbandster2smac.convert([self.path_to_result_mixed_categorical_json])
+            hpbandster2smac.convert([self.path_to_result_mixed_categorical_pcs])
             # Missing configfile
-            self.assertRaises(ValueError, hpbandster2smac.convert, self.path_to_result_mixed_categorical_missing)
+            self.assertRaises(ValueError, hpbandster2smac.convert, [self.path_to_result_mixed_categorical_missing])
         except ImportError:
             pass